rename depiction controls to clipping plane controls
