+++ finn/_qt/layer_controls/qt_layer_clipping_plane_controls.py
